Day 6
Due to yesterday's challenge, I know that iterating through each possible option was likely going to not work, so I looked at the problem and noticed that the graph looked like a parabola, so I tested an equation in Desmos and got it to work.

Because of my optimization for Part 1, Part 2 was a breeze, as I only needed to change the parser.