Day 10 + Djikstra
Part 1 was fairly simple, but Part 2 threw me for a loop. I had the original idea of checking to see which points has a path to the edge, but I was told that I was over complicating things, so I brainstormed and thought of another solution, where I check to see which points are in the polygon formed by the path, and after a lot of debugging, I got it to work!

I've heard a lot of people mention the Djikstra algorithm is often a challenge in Advent of Code, so I implemented it for when I'd need it.